Ahora se comprueba cuando inicia la app, quienes deben estar activos e inactivos, y al pagar deudas cambia el estado de activo e inactivo.

+++ asociacion/src/main/java/com/asociacion/models/Fee.java
@@ -23,6 +23,6 @@ public class Fee {
 
     @Temporal(TemporalType.DATE)
     private Date date; // "2024-11-12"
-    private Long year;
+    private int year;
 
 }
